chore(node:stream): extract node_stream test module to sibling file (2000-line gate)
The #1746/#1537 additions pushed node_stream.rs to 2043 lines, over the
file-size CI gate. Move the #[cfg(test)] mod tests into node_stream_tests.rs
via #[path] (tests use `use super::*`, so no visibility changes); node_stream.rs
drops to 1841 lines.
